Migrated selected courses from cards to list. Added CourseListing component which displays course information and several actions. TODO: use centralized state management to stop using v-on:$listeners event bubble hack and possibly fix debilitating rerendering
